feat(settings): move "Feedback senden" from BottomNav to Settings header (#90)
BottomNav now contains only main views (Transkriptionen, Einstellungen).
The Feedback action moves to the Settings home header as a primary
button matching the existing "+ Eintrag hinzufügen" pattern, hidden on
sub-pages so contextual actions stay unique. Underlying
window.api.feedback.send() is unchanged.
