+++ src/renderer/src/views/Settings.tsx
@@ -8,6 +8,7 @@ import ModelsSettings from '../components/settings/ModelsSettings'
 import { useTheme } from '../hooks/useTheme'
 import { useAppUpdate } from '../hooks/useAppUpdate'
 import { useModelUpdates } from '../hooks/useModelUpdates'
+import { useToast } from '../hooks/useToast'
 
 type SubPage = 'sperrliste' | 'darstellung' | 'modelle' | 'ueber'
 
@@ -23,9 +24,33 @@ const THEME_LABELS = { light: 'Hell', dark: 'Dunkel', system: 'System' } as cons
 export default function Settings(): React.JSX.Element {
   const [subpage, setSubpage] = useState<SubPage | null>(null)
   const blocklistRef = useRef<BlocklistManagerHandle>(null)
+  const toast = useToast()
 
-  const headerAction =
-    subpage === 'sperrliste' ? (
+  const handleSendFeedback = async (): Promise<void> => {
+    try {
+      await window.api.feedback.send()
+      toast.success(
+        'Feedback-Mail vorbereitet — Inhalt wurde zusätzlich in die Zwischenablage kopiert.'
+      )
+    } catch (error) {
+      console.error('[feedback] send failed:', error)
+      toast.error('Feedback konnte nicht vorbereitet werden.')
+    }
+  }
+
+  let headerAction: React.ReactNode = null
+  if (subpage === null) {
+    headerAction = (
+      <button
+        type="button"
+        className="titlebar-no-drag rounded-lg bg-primary px-4 py-2 text-sm font-semibold text-white transition-colors hover:bg-primary-hover"
+        onClick={handleSendFeedback}
+      >
+        Feedback senden
+      </button>
+    )
+  } else if (subpage === 'sperrliste') {
+    headerAction = (
       <button
         type="button"
         className="titlebar-no-drag rounded-lg bg-primary px-4 py-2 text-sm font-semibold text-white transition-colors hover:bg-primary-hover"
@@ -33,7 +58,8 @@ export default function Settings(): React.JSX.Element {
       >
         + Eintrag hinzufügen
       </button>
-    ) : null
+    )
+  }
 
   return (
     <div className="flex min-h-0 flex-1 flex-col">
